% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculadora.R
\name{dividir}
\alias{dividir}
\title{Function to divide}
\usage{
dividir(a, b)
}
\arguments{
\item{a}{A number}

\item{b}{It is another number}
}
\description{
This function is good to divide
}
\examples{
dividir(3,2)

}
