% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculadora.R
\name{multiplicar}
\alias{multiplicar}
\title{Function to multiply}
\usage{
multiplicar(a, b)
}
\arguments{
\item{a}{A number}

\item{b}{It is another number}
}
\description{
This function is good to multiply
}
\examples{
multiplicar(3,2)

}
